show deleted text on hover
